+++ alphabase/protein/fasta.py
@@ -473,7 +473,10 @@ def parse_labels(labels:list):
     return label_aas, label_mod_dict, nterm_label_mod, cterm_label_mod
         
 def create_labeling_peptide_df(peptide_df:pd.DataFrame, labels:list):
+    if len(peptide_df) == 0: return peptide_df
+
     df = peptide_df.copy()
+
     (
         label_aas, label_mod_dict, 
         nterm_label_mod, cterm_label_mod
@@ -499,7 +502,8 @@ def protein_idxes_to_names(protein_idxes:str, protein_names:list):
     proteins = [protein for protein in proteins if protein]
     return ';'.join(proteins)
 
-def append_special_modifications(df:pd.DataFrame, 
+def append_special_modifications(
+    df:pd.DataFrame, 
     var_mods:list = ['Phospho@S','Phospho@T','Phospho@Y'], 
     min_mod_num:int=0, max_mod_num:int=1, 
     max_peptidoform_num:int=100,
@@ -545,7 +549,7 @@ def append_special_modifications(df:pd.DataFrame,
     pd.DataFrame
         The precursor_df with new modification added.
     """
-    if len(var_mods) == 0: 
+    if len(var_mods) == 0 or len(df) == 0: 
         return df
 
     if cannot_modify_pep_nterm_aa:
@@ -643,7 +647,8 @@ class SpecLibFasta(SpecLibBase):
         special_mods_cannot_modify_pep_n_term:bool=False,
         special_mods_cannot_modify_pep_c_term:bool=False,
         decoy: str = None, # or pseudo_reverse or diann
-        I_to_L=False,
+        include_contaminants:bool=False,
+        I_to_L:bool=False,
     ):
         """
         Parameters
@@ -725,6 +730,9 @@ class SpecLibFasta(SpecLibBase):
         decoy : str, optional
             Decoy type, see `alphabase.spectral_library.decoy_library`,
             by default None
+
+        include_contaminants : bool, optional
+            If include contaminants.fasta, by default False
         """
         super().__init__(
             charged_frag_types=charged_frag_types,
@@ -734,6 +742,7 @@ class SpecLibFasta(SpecLibBase):
         )
         self.protein_df:pd.DataFrame() = pd.DataFrame()
         self.I_to_L = I_to_L
+        self.include_contaminants = include_contaminants
         self.max_peptidoform_num = 100
         self._digest = Digest(
             protease, max_missed_cleavages,
@@ -851,7 +860,9 @@ class SpecLibFasta(SpecLibBase):
                 mod_set.add(mod[-1])
         return False
 
-    def import_and_process_fasta(self, fasta_files:Union[str,list]):
+    def import_and_process_fasta(self, 
+        fasta_files:list,
+    ):
         """
         Import and process a fasta file or a list of fasta files.
         It includes 3 steps:
@@ -862,9 +873,13 @@ class SpecLibFasta(SpecLibBase):
 
         Parameters
         ----------
-        fasta_files : Union[str,list]
+        fasta_files : list
             A fasta file or a list of fasta files
         """
+        if self.include_contaminants:
+            fasta_files.append(os.path.join(
+                CONST_FILE_FOLDER, 'contaminants.fasta'
+            ))
         protein_dict = load_all_proteins(fasta_files)
         self.import_and_process_protein_dict(protein_dict)
 
@@ -942,6 +957,10 @@ class SpecLibFasta(SpecLibBase):
         fasta_files : list
             fasta file list
         """
+        if self.include_contaminants:
+            fasta_files.append(os.path.join(
+                CONST_FILE_FOLDER, 'contaminants.fasta'
+            ))
         protein_dict = load_all_proteins(fasta_files)
         self.get_peptides_from_protein_dict(protein_dict)
 
@@ -1130,6 +1149,11 @@ class SpecLibFasta(SpecLibBase):
         if 'is_prot_cterm' not in self._precursor_df.columns:
             self._precursor_df['is_prot_cterm'] = False
         
+        if len(self._precursor_df) == 0:
+            self._precursor_df['mods'] = ""
+            self._precursor_df['mod_sites'] = ""
+            return
+            
         (
             self._precursor_df['mods'],
             self._precursor_df['mod_sites']