+++ HISTORY.md
@@ -4,6 +4,14 @@ Follow the changelog format from https://keepachangelog.com/en/1.0.0/.
 
 ## 1.0.0 - TODO Next Release
 
+### Added
+
+- `Percolator` and `SupervisedPercolator` (`alphabase.scoring.ml_scoring.py`).
+- `SpecLibFat` for flat fragment dataframes from `SpecLibBase` (`alphabase.spectral_library.flat.py`).
+- `SpecLibBase` from TSV library (`alphabase.spectral_library.reader.py`)
+- `LOESSRegression` for recalibration stuffs (`alphabase.statistics.regression.py`)
+- Auto include contaminants.fasta with `include_contaminants` in `alphabase.protein.fasta.SpecLibFasta`.
+
 ## 0.4.0 - 2022.12.28
 
 ### Changed
@@ -15,12 +23,9 @@ Follow the changelog format from https://keepachangelog.com/en/1.0.0/.
 ### Added
 
 - `alphabase.protein.fasta.SpecLibFasta.add_peptide_labeling()` supports mDIA.
-- `alphabase.spectral_library.flat.SpecLibFlat` class for flat spec_lib.
 - Move `alphabase.spectral_library.translate` module (.py) from alphapeptdeep to alphabase.
 - `alphabase.constants.modification.add_new_modifications()` to add user-defined modifications that are not in UniMod.
 - `alphabase.psm_reader.psm_reader.PSMReaderBase.add_modification_mapping()` to support arbitrary modifications of different search engines.
-- `alphabase.statistics.regression.LOESSRegression`
-- `alphabase.scoring` module added but not finished yet.
 
 ### Changed